requirements.txt
cmake
face-recognition
streamlit
opencv-python-headless
Pillow
numpy